requirements_updated.txt
# Dependencias originales
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
qrcode[pil]
Pillow
python-barcode 
reportlab==4.4.2

# Nuevas dependencias para mejoras
bcrypt==4.1.2
PyJWT==2.8.0
psutil==5.9.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
schedule==1.2.0

# Para desarrollo y testing
black==23.11.0
flake8==6.1.0
mypy==1.7.1